Refactor GUI into dedicated package
