+++ gui/__init__.py
@@ -1,0 +1,3 @@
+from .main_window import MainWindow, main
+
+__all__ = ["MainWindow", "main"]
